gracefully shut down the channel on drop instead of discarding events
Previously, if the debouncer's shutdown flag was set while there was
anything in its accumulator, the contents of the accumulator would never
be sent through the mpsc channel, effectively causing the events folded
into the accumulator to be lost. This patch modifies the shutdown
behaviour to send the contents of the accumulator through the channel
before stopping the debouncer and closing the channel.